requirements.txt
fastapi
pydantic
numpy
matplotlib
Pillow
uvicorn[standard]